requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pytest
groq
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2